requirements.txt
dspy-ai
python_dotenv
